checkpoint: moves "/Base64/" to "/Base64_dep/"
