+++ src/library/Sequence/Byte/Encoded/Base64_dep/ParsingError.ts
