+++ src/library/Sequence/Byte/Encoded/Base64_dep/definition.declared.ts
