+++ src/library/Sequence/Byte/Encoded/definition.assembled.members.ts
@@ -4,4 +4,4 @@ export {
 
 export {
   Sequence_Byte_Encoded_Base64 as Base64,
-} from './Base64';
+} from './Base64_dep';
